refactor(capture): resolve os file dependency on plugin.capture
