+++ src/plugin/capture/capture.js
@@ -1,38 +1,19 @@
 goog.module('plugin.capture');
 goog.module.declareLegacyNamespace();
 
-const MapContainer = goog.require('os.MapContainer');
-const osMap = goog.require('os.map');
-
 /**
  * Get the map canvas element.
  * @return {HTMLCanvasElement} The map canvas element
+ * @deprecated Please use os.capture.getMapCanvas instead.
  */
-const getMapCanvas = function() {
-  var mapCanvas;
-  if (MapContainer.getInstance().is3DEnabled()) {
-    mapCanvas = /** @type {HTMLCanvasElement} */ (document.querySelector(osMap.WEBGL_CANVAS));
-  } else {
-    mapCanvas = /** @type {HTMLCanvasElement} */ (document.querySelector(osMap.OPENLAYERS_CANVAS));
-  }
-
-  return mapCanvas || null;
-};
-
+const getMapCanvas = getMapCanvas;
 
 /**
  * Get the map canvas pixel ratio.
  * @return {number} The map canvas pixel ratio.
+ * @deprecated Please use os.capture.getMapPixelRatio instead.
  */
-const getMapPixelRatio = function() {
-  var mapCanvas = getMapCanvas();
-  if (mapCanvas) {
-    var mapRect = mapCanvas.getBoundingClientRect();
-    return mapCanvas.width / mapRect.width;
-  }
-
-  return 1;
-};
+const getMapPixelRatio = getMapPixelRatio;
 
 exports = {
   getMapCanvas,
